+++ src/filetransfer/scp_transfer.rs
@@ -37,9 +37,9 @@ use crate::utils::parser::parse_lstime;
 use regex::Regex;
 use ssh2::{Channel, Session};
 use std::io::{BufReader, BufWriter, Read, Write};
-use std::net::TcpStream;
+use std::net::{SocketAddr, TcpStream, ToSocketAddrs};
 use std::path::{Path, PathBuf};
-use std::time::SystemTime;
+use std::time::{Duration, SystemTime};
 
 /// ## ScpFileTransfer
 ///
@@ -311,12 +311,34 @@ impl FileTransfer for ScpFileTransfer {
         password: Option<String>,
     ) -> Result<Option<String>, FileTransferError> {
         // Setup tcp stream
-        let tcp: TcpStream = match TcpStream::connect(format!("{}:{}", address, port)) {
-            Ok(stream) => stream,
-            Err(err) => {
+        let socket_addresses: Vec<SocketAddr> =
+            match format!("{}:{}", address, port).to_socket_addrs() {
+                Ok(s) => s.collect(),
+                Err(err) => {
+                    return Err(FileTransferError::new_ex(
+                        FileTransferErrorType::BadAddress,
+                        format!("{}", err),
+                    ))
+                }
+            };
+        let mut tcp: Option<TcpStream> = None;
+        // Try addresses
+        for socket_addr in socket_addresses.iter() {
+            match TcpStream::connect_timeout(&socket_addr, Duration::from_secs(30)) {
+                Ok(stream) => {
+                    tcp = Some(stream);
+                    break;
+                }
+                Err(_) => continue,
+            }
+        }
+        // If stream is None, return connection timeout
+        let tcp: TcpStream = match tcp {
+            Some(t) => t,
+            None => {
                 return Err(FileTransferError::new_ex(
-                    FileTransferErrorType::BadAddress,
-                    format!("{}", err),
+                    FileTransferErrorType::ConnectionError,
+                    String::from("Connection timeout"),
                 ))
             }
         };